Move colorToVector3 method from renderingUtil.ts to colorConverter.ts
